Moves to firehose package
