+++ brave/src/main/java/brave/RealScopedSpan.java
@@ -1,7 +1,7 @@
 package brave;
 
-import brave.internal.recorder.Firehose;
-import brave.internal.recorder.MutableSpan;
+import brave.firehose.Firehose;
+import brave.firehose.MutableSpan;
 import brave.internal.recorder.PendingSpans;
 import brave.propagation.CurrentTraceContext.Scope;
 import brave.propagation.TraceContext;
