+++ brave/src/main/java/brave/RealSpan.java
@@ -1,7 +1,7 @@
 package brave;
 
-import brave.internal.recorder.Firehose;
-import brave.internal.recorder.MutableSpan;
+import brave.firehose.Firehose;
+import brave.firehose.MutableSpan;
 import brave.internal.recorder.PendingSpans;
 import brave.propagation.TraceContext;
