+++ brave/src/main/java/brave/Tracer.java
@@ -3,8 +3,8 @@ package brave;
 import brave.internal.InternalPropagation;
 import brave.internal.Nullable;
 import brave.internal.Platform;
-import brave.internal.recorder.Firehose;
-import brave.internal.recorder.MutableSpan;
+import brave.firehose.Firehose;
+import brave.firehose.MutableSpan;
 import brave.internal.recorder.PendingSpan;
 import brave.internal.recorder.PendingSpans;
 import brave.propagation.CurrentTraceContext;
