+++ brave/src/main/java/brave/internal/recorder/PendingSpan.java
@@ -1,6 +1,7 @@
 package brave.internal.recorder;
 
 import brave.Clock;
+import brave.firehose.MutableSpan;
 
 public final class PendingSpan {
   final MutableSpan state;
