+++ brave/src/test/java/brave/internal/recorder/MutableSpanTest.java
@@ -1,5 +1,6 @@
 package brave.internal.recorder;
 
+import brave.firehose.MutableSpan;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import org.junit.Test;
