+++ instrumentation/benchmarks/src/main/java/brave/internal/recorder/MutableSpanBenchmarks.java
@@ -14,6 +14,7 @@
 package brave.internal.recorder;
 
 import brave.Span;
+import brave.firehose.MutableSpan;
 import java.util.concurrent.TimeUnit;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
