+++ spring-beans/src/main/java/brave/spring/beans/TracingFactoryBean.java
@@ -3,7 +3,7 @@ package brave.spring.beans;
 import brave.Clock;
 import brave.ErrorParser;
 import brave.Tracing;
-import brave.internal.recorder.Firehose;
+import brave.firehose.Firehose;
 import brave.propagation.CurrentTraceContext;
 import brave.propagation.Propagation;
 import brave.sampler.Sampler;
