+++ spring-beans/src/test/java/brave/spring/beans/TracingFactoryBeanTest.java
@@ -3,7 +3,7 @@ package brave.spring.beans;
 import brave.Clock;
 import brave.ErrorParser;
 import brave.Tracing;
-import brave.internal.recorder.Firehose;
+import brave.firehose.Firehose;
 import brave.propagation.ExtraFieldPropagation;
 import brave.propagation.ThreadLocalCurrentTraceContext;
 import brave.sampler.Sampler;
